requirements.txt
pandas
numpy
pillow
matplotlib 
Flask 
fastapi
uvicorn
jinja2 
python-multipart